fix win64 build, issue #109

+++ deploy.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 # Build and deploy plugin
 #
-# Works on Linux, FreeBSD and Windows
+# Works on Linux, FreeBSD and Windows(MSYS)
 #
 # Options:
 #
@@ -58,38 +58,8 @@ else
 fi
 GMAGICK_URL=$GMAGICK_URL/GraphicsMagick-$GMAGICK.tar.gz
 
-ZLIB=1.2.8
-ZLIB_URL=http://prdownloads.sourceforge.net/libpng/zlib-${ZLIB}.tar.gz?download
-
 PNG=1.2.52
 PNG_URL=http://prdownloads.sourceforge.net/libpng/libpng-${PNG}.tar.gz?download
-
-EXPAT=2.1.0
-EXPAT_URL=http://sourceforge.net/projects/expat/files/expat/${EXPAT}/expat-${EXPAT}.tar.gz/download
-
-FCONFIG=2.10.2
-FCONFIG_URL=http://www.freedesktop.org/software/fontconfig/release/fontconfig-${FCONFIG}.tar.gz
-
-FTYPE=2.4.11
-FTYPE_URL=http://sourceforge.net/projects/freetype/files/freetype2/${FTYPE}/freetype-${FTYPE}.tar.gz/download
-
-FFI=3.2.1
-FFI_URL=ftp://sourceware.org/pub/libffi/libffi-$FFI.tar.gz
-
-GLIB=2.42.2
-GLIB_URL=http://ftp.acc.umu.se/pub/GNOME/sources/glib/2.42/glib-$GLIB.tar.xz
-
-BUZZ=0.9.40
-BUZZ_URL=http://www.freedesktop.org/software/harfbuzz/release/harfbuzz-$BUZZ.tar.bz2
-
-PIX=0.32.6
-PIX_URL=http://cairographics.org/releases/pixman-$PIX.tar.gz
-
-CAIRO=1.14.2
-CAIRO_URL=http://cairographics.org/releases/cairo-$CAIRO.tar.xz
-
-PANGO=1.37.0
-PANGO_URL=http://ftp.gnome.org/pub/GNOME/sources/pango/1.37/pango-$PANGO.tar.xz
 
 if [ -z "$VERSION" ]; then
   ARENA=0.8
@@ -170,30 +140,8 @@ if [ ! -d $PREFIX ]; then
   (cd $PREFIX ; ln -sf lib lib64)
 fi
 
-# zlib
-if [ ! -f ${PREFIX}/lib/libz.a ] && [ "$OS" == "Msys" ]; then
-  if [ ! -f $CWD/3rdparty/zlib-$ZLIB.tar.gz ]; then
-    wget $ZLIB_URL -O $CWD/3rdparty/zlib-$ZLIB.tar.gz || exit 1
-  fi
-  if [ ! -d $CWD/3rdparty/zlib-$ZLIB ]; then
-    tar xvf $CWD/3rdparty/zlib-$ZLIB.tar.gz -C $CWD/3rdparty/ || exit 1
-  fi
-  cd $CWD/3rdparty/zlib-$ZLIB || exit 1
-  if [ -f $CWD/3rdparty/Makefile.zlib ] && [ "$BIT" == "64" ]; then
-    cat $CWD/3rdparty/Makefile.zlib > win32/Makefile.gcc || exit 1
-  fi
-  make distclean
-  CFLAGS="-m${BIT} ${BF}" CXXFLAGS="-m${BIT} ${BF} -I${PREFIX}/include" CPPFLAGS="-I${PREFIX}/include -L${PREFIX}/lib" make -f win32/Makefile.gcc || exit 1
-  cp libz.a ${PREFIX}/lib/
-  cp *.h ${PREFIX}/include/
-  mkdir -p $PREFIX/share/doc/zlib/ || exit 1
-  cp README $PREFIX/share/doc/zlib/ || exit 1
-  cd .. || exit 1
-  rm -rf zlib-$ZLIB || exit 1
-fi
-
 # libpng
-if [ ! -f ${PREFIX}/lib/libpng.a ]; then
+if [ ! -f ${PREFIX}/lib/libpng.a ] && [ "$OS" != "Msys" ]; then
   if [ ! -f $CWD/3rdparty/libpng-$PNG.tar.gz ]; then
     wget $PNG_URL -O $CWD/3rdparty/libpng-$PNG.tar.gz || exit 1
   fi
@@ -209,183 +157,6 @@ if [ ! -f ${PREFIX}/lib/libpng.a ]; then
   cd .. || exit 1
   rm -rf libpng-$PNG || exit 1
 fi
-
-# expat
-if [ ! -f ${PREFIX}/lib/libexpat.a ] && [ "$OS" == "Msys" ]; then
-  if [ ! -f $CWD/3rdparty/expat-$EXPAT.tar.gz ]; then
-    wget $EXPAT_URL -O $CWD/3rdparty/expat-$EXPAT.tar.gz || exit 1
-  fi
-  if [ ! -d $CWD/3rdparty/expat-$EXPAT ]; then
-    tar xvf $CWD/3rdparty/expat-$EXPAT.tar.gz -C $CWD/3rdparty/ || exit 1
-  fi
-  cd $CWD/3rdparty/expat-$EXPAT || exit 1
-  make distclean
-  CFLAGS="-m${BIT} ${BF}" CXXFLAGS="-m${BIT} ${BF} -I${PREFIX}/include" CPPFLAGS="-I${PREFIX}/include -L${PREFIX}/lib" ./configure --libdir=${PREFIX}/lib --prefix=${PREFIX} --enable-static --disable-shared || exit 1
-  make -j$JOBS install || exit 1
-  mkdir -p $PREFIX/share/doc/expat/ || exit 1
-  cp COPYING $PREFIX/share/doc/expat/ || exit 1
-  cd .. || exit 1
-  rm -rf expat-$EXPAT || exit 1
-fi
-
-# freetype
-if [ ! -f ${PREFIX}/lib/libfreetype.a ] && [ "$OS" == "Msys" ]; then
-  if [ ! -f $CWD/3rdparty/freetype-$FTYPE.tar.gz ]; then
-    wget $FTYPE_URL -O $CWD/3rdparty/freetype-$FTYPE.tar.gz || exit 1
-  fi
-  if [ ! -d $CWD/3rdparty/freetype-$FTYPE ]; then
-    tar xvf $CWD/3rdparty/freetype-$FTYPE.tar.gz -C $CWD/3rdparty/ || exit 1
-  fi
-  cd $CWD/3rdparty/freetype-$FTYPE || exit 1
-  make distclean
-  CFLAGS="-m${BIT} ${BF}" CXXFLAGS="-m${BIT} ${BF} -I${PREFIX}/include" CPPFLAGS="-I${PREFIX}/include -L${PREFIX}/lib -lexpat -lz" ./configure --libdir=${PREFIX}/lib --prefix=${PREFIX} --enable-static --disable-shared || exit 1
-  make -j$JOBS install || exit 1
-  mkdir -p $PREFIX/share/doc/freetype/ || exit 1
-  cp README $PREFIX/share/doc/freetype/ || exit 1
-  cd .. || exit 1
-  rm -rf freetype-$FTYPE || exit 1
-fi
-
-# fontconfig
-if [ ! -f ${PREFIX}/lib/libfontconfig.a ] && [ "$OS" == "Msys" ]; then
-  if [ ! -f $CWD/3rdparty/fontconfig-$FCONFIG.tar.gz ]; then
-    wget $FCONFIG_URL -O $CWD/3rdparty/fontconfig-$FCONFIG.tar.gz || exit 1
-  fi
-  if [ ! -d $CWD/3rdparty/fontconfig-$FCONFIG ]; then
-    tar xvf $CWD/3rdparty/fontconfig-$FCONFIG.tar.gz -C $CWD/3rdparty/ || exit 1
-  fi
-  cd $CWD/3rdparty/fontconfig-$FCONFIG || exit 1
-  make distclean
-  CFLAGS="-m${BIT} ${BF}" CXXFLAGS="-m${BIT} ${BF} -I${PREFIX}/include" CPPFLAGS="-I${PREFIX}/include -L${PREFIX}/lib -lfreetype -lexpat -lz" ./configure --bindir=${PREFIX} --libdir=${PREFIX}/lib --prefix=${PREFIX} --disable-docs --enable-static --disable-shared || exit 1
-  make -j$JOBS install || exit 1
-  mkdir -p $PREFIX/share/doc/fontconfig/ || exit 1
-  cp COPYING $PREFIX/share/doc/fontconfig/ || exit 1
-  cd .. || exit 1
-  rm -rf fontconfig-$FCONFIG || exit 1
-fi
-
-# libffi
-#if [ ! -f ${PREFIX}/lib/pkgconfig/libffi.pc ]; then
-#  if [ ! -f $CWD/3rdparty/libffi-$FFI.tar.gz ]; then
-#    wget $FFI_URL -O $CWD/3rdparty/libffi-$FFI.tar.gz || exit 1
-#  fi
-#  if [ ! -d $CWD/3rdparty/libffi-$FFI ]; then
-#    tar xvf $CWD/3rdparty/libffi-$FFI.tar.gz -C $CWD/3rdparty/ || exit 1
-#  fi
-#  cd $CWD/3rdparty/libffi-$FFI || exit 1
-#  make distclean
-#  CFLAGS="-m${BIT} ${BF}" CXXFLAGS="-m${BIT} ${BF} -I${PREFIX}/include" CPPFLAGS="-I${PREFIX}/include -L${PREFIX}/lib -lfreetype -lexpat -lz" ./configure --bindir=${PREFIX} --libdir=${PREFIX}/lib --prefix=${PREFIX} --disable-docs --enable-static --disable-shared || exit 1
-#  make -j$JOBS install || exit 1
-#  mkdir -p $PREFIX/share/doc/libffi/ || exit 1
-#  cp COPYING* LIC* $PREFIX/share/doc/libffi/
-#  cd .. || exit 1
-#  rm -rf libffi-$FFI || exit 1
-#fi
-
-# glib
-#if [ ! -f ${PREFIX}/lib/pkgconfig/glib-2.0.pc ]; then
-#  if [ ! -f $CWD/3rdparty/glib-$GLIB.tar.xz ]; then
-#    wget $GLIB_URL -O $CWD/3rdparty/glib-$GLIB.tar.xz || exit 1
-#  fi
-#  if [ ! -d $CWD/3rdparty/glib-$GLIB ]; then
-#    tar xvf $CWD/3rdparty/glib-$GLIB.tar.xz -C $CWD/3rdparty/ || exit 1
-#  fi
-#  cd $CWD/3rdparty/glib-$GLIB || exit 1
-#  make distclean
-#  CFLAGS="-m${BIT} ${BF}" CXXFLAGS="-m${BIT} ${BF} -I${PREFIX}/include" CPPFLAGS="-I${PREFIX}/include -L${PREFIX}/lib -lfreetype -lexpat -lz" ./configure --bindir=${PREFIX} --libdir=${PREFIX}/lib --prefix=${PREFIX} --disable-docs --disable-static --enable-shared || exit 1
-#  make -j$JOBS install || exit 1
-#  mkdir -p $PREFIX/share/doc/glib/ || exit 1
-#  cp COPYING* LIC* $PREFIX/share/doc/glib/
-#  cd .. || exit 1
-#  rm -rf glib-$GLIB || exit 1
-#fi
-
-# pixman
-#if [ ! -f ${PREFIX}/lib/pkgconfig/pixman-1.pc ]; then
-#  if [ ! -f $CWD/3rdparty/pixman-$PIX.tar.gz ]; then
-#    wget $PIX_URL -O $CWD/3rdparty/pixman-$PIX.tar.gz || exit 1
-#  fi
-#  if [ ! -d $CWD/3rdparty/pixman-$PIX ]; then
-#    tar xvf $CWD/3rdparty/pixman-$PIX.tar.gz -C $CWD/3rdparty/ || exit 1
-#  fi
-#  cd $CWD/3rdparty/pixman-$PIX || exit 1
-#  make distclean
-#  CFLAGS="-m${BIT} ${BF}" CXXFLAGS="-m${BIT} ${BF} -I${PREFIX}/include" CPPFLAGS="-I${PREFIX}/include -L${PREFIX}/lib" ./configure --bindir=${PREFIX} --libdir=${PREFIX}/lib --prefix=${PREFIX} --disable-docs --enable-static --disable-shared || exit 1
-#  make -j$JOBS install || exit 1
-#  mkdir -p $PREFIX/share/doc/pixman/ || exit 1
-#  cp COPYING* LIC* $PREFIX/share/doc/pixman/
-#  cd .. || exit 1
-#  rm -rf pixman-$PIX || exit 1
-#fi
-
-# cairo
-#if [ ! -f ${PREFIX}/lib/pkgconfig/cairo.pc ]; then
-#  if [ ! -f $CWD/3rdparty/cairo-$CAIRO.tar.xz ]; then
-#    wget $CAIRO_URL -O $CWD/3rdparty/cairo-$CAIRO.tar.xz || exit 1
-#  fi
-#  if [ ! -d $CWD/3rdparty/cairo-$CAIRO ]; then
-#    tar xvf $CWD/3rdparty/cairo-$CAIRO.tar.xz -C $CWD/3rdparty/ || exit 1
-#  fi
-#  cd $CWD/3rdparty/cairo-$CAIRO || exit 1
-#  make distclean
-#  CFLAGS="-m${BIT} ${BF}" CXXFLAGS="-m${BIT} ${BF} -I${PREFIX}/include -I${PREFIX}/include/pixman-1" CPPFLAGS="-I${PREFIX}/include -L${PREFIX}/lib -lpixman-1" ./configure --bindir=${PREFIX} --libdir=${PREFIX}/lib --prefix=${PREFIX} --disable-docs --disable-shared --enable-static || exit 1
-#  make -j$JOBS install || exit 1
-#  mkdir -p $PREFIX/share/doc/cairo/ || exit 1
-#  cp COPYING* LIC* $PREFIX/share/doc/cairo/
-#  cd .. || exit 1
-#  rm -rf cairo-$CAIRO || exit 1
-#fi
-
-# harfbuzz
-#if [ ! -f ${PREFIX}/lib/pkgconfig/harfbuzz.pc ]; then
-#  if [ ! -f $CWD/3rdparty/harfbuzz-$BUZZ.tar.bz2 ]; then
-#    wget $BUZZ_URL -O $CWD/3rdparty/harfbuzz-$BUZZ.tar.bz2 || exit 1
-#  fi
-#  if [ ! -d $CWD/3rdparty/harfbuzz-$BUZZ ]; then
-#    tar xvf $CWD/3rdparty/harfbuzz-$BUZZ.tar.bz2 -C $CWD/3rdparty/ || exit 1
-#  fi
-#  cd $CWD/3rdparty/harfbuzz-$BUZZ || exit 1
-#  make distclean
-#  CFLAGS="-m${BIT} ${BF}" CXXFLAGS="-m${BIT} ${BF} -I${PREFIX}/include" CPPFLAGS="-I${PREFIX}/include -L${PREFIX}/lib -lfreetype -lexpat -lz" ./configure --bindir=${PREFIX} --libdir=${PREFIX}/lib --prefix=${PREFIX} --disable-docs --enable-static --disable-shared --with-freetype --with-cairo --with-gobject --with-glib --without-icu || exit 1
-#  make -j$JOBS install || exit 1
-#  mkdir -p $PREFIX/share/doc/harfbuzz/ || exit 1
-#  cp COPYING* LIC* $PREFIX/share/doc/harfbuzz/
-#  cd .. || exit 1
-#  rm -rf harfbuzz-$BUZZ || exit 1
-  # rebuild freetype
-#  if [ ! -f $CWD/3rdparty/freetype-$FTYPE.tar.gz ]; then
-#    wget $FTYPE_URL -O $CWD/3rdparty/freetype-$FTYPE.tar.gz || exit 1
-#  fi
-#  if [ ! -d $CWD/3rdparty/freetype-$FTYPE ]; then
-#    tar xvf $CWD/3rdparty/freetype-$FTYPE.tar.gz -C $CWD/3rdparty/ || exit 1
-#  fi
-#  cd $CWD/3rdparty/freetype-$FTYPE || exit 1
-#  make distclean
-#  CFLAGS="-m${BIT} ${BF}" CXXFLAGS="-m${BIT} ${BF} -I${PREFIX}/include" CPPFLAGS="-I${PREFIX}/include -L${PREFIX}/lib -lexpat -lz" ./configure --libdir=${PREFIX}/lib --prefix=${PREFIX} --enable-static --disable-shared || exit 1
-#  make -j$JOBS install || exit 1
-#  mkdir -p $PREFIX/share/doc/freetype/ || exit 1
-#  cp README $PREFIX/share/doc/freetype/ || exit 1
-#  cd .. || exit 1
-#  rm -rf freetype-$FTYPE || exit 1
-#fi
-
-# pango
-#if [ ! -f ${PREFIX}/lib/pkgconfig/pango.pc ]; then
-#  if [ ! -f $CWD/3rdparty/pango-$PANGO.tar.xz ]; then
-#    wget $PANGO_URL -O $CWD/3rdparty/pango-$PANGO.tar.xz || exit 1
-#  fi
-#  if [ ! -d $CWD/3rdparty/pango-$PANGO ]; then
-#    tar xvf $CWD/3rdparty/pango-$PANGO.tar.xz -C $CWD/3rdparty/ || exit 1
-#  fi
-#  cd $CWD/3rdparty/pango-$PANGO || exit 1
-#  make distclean
-#  FONTCONFIG_CFLAGS="-I${PREFIX}/include" FONTCONFIG_LIBS="-L${PREFIX}/lib -lfontconfig" CFLAGS="-m${BIT} ${BF}" CXXFLAGS="-m${BIT} ${BF} -I${PREFIX}/include" CPPFLAGS="-I${PREFIX}/include -L${PREFIX}/lib -lfontconfig -lfreetype -lexpat -lz" ./configure --bindir=${PREFIX} --libdir=${PREFIX}/lib --prefix=${PREFIX} --disable-docs --disable-static --enable-shared --with-included-modules=basic-fc || exit 1
-#  make -j$JOBS install || exit 1
-#  mkdir -p $PREFIX/share/doc/pango/ || exit 1
-#  cp COPYING* LIC* $PREFIX/share/doc/pango/
-#  cd .. || exit 1
-#  rm -rf pango-$PANGO || exit 1
-#fi
 
 # magick
 if [ "$CLEAN" == "1" ]; then
@@ -452,8 +223,8 @@ if [ "$OS" != "Msys" ]; then
   $MAKE USE_SVG=1 OPENCL=$USE_CL USE_PANGO=1 STATIC=1 FREEBSD=$USE_FREEBSD BITS=$BIT CONFIG=$TAG clean
   $MAKE USE_SVG=1 OPENCL=$USE_CL USE_PANGO=1 STATIC=1 FREEBSD=$USE_FREEBSD BITS=$BIT CONFIG=$TAG || exit 1
 else
-  make STATIC=1 OPENCL=$USE_CL MINGW=1 BIT=$BIT CONFIG=release clean
-  make STATIC=1 OPENCL=$USE_CL MINGW=1 BIT=$BIT CONFIG=release || exit 1
+  make STATIC=1 USE_SVG=1 OPENCL=$USE_CL USE_PANGO=1 MINGW=1 BIT=$BIT CONFIG=release clean
+  make STATIC=1 USE_SVG=1 OPENCL=$USE_CL USE_PANGO=1 MINGW=1 BIT=$BIT CONFIG=release || exit 1
 fi
 
 cd $CWD || exit 1
@@ -464,8 +235,10 @@ mkdir $CWD/$PKG || exit 1
 cp LICENSE README.md $CWD/$PKG/ || exit 1
 cp $PREFIX/share/doc/ImageMagick/LICENSE $CWD/$PKG/LICENSE.ImageMagick || exit 1
 cp OpenFX/Support/LICENSE $CWD/$PKG/LICENSE.OpenFX || exit 1
-cp $PREFIX/share/doc/libpng/LICENSE $CWD/$PKG/LICENSE.libpng || exit 1
-
+cp OpenFX-IO/LICENSE $CWD/$PKG/LICENSE.OpenFX-IO || exit 1
+if [ "$OS" != "Msys" ]; then 
+  cp $PREFIX/share/doc/libpng/LICENSE $CWD/$PKG/LICENSE.libpng || exit 1
+fi
 
 if [ "$CL" == "1" ]; then
   cp $CWD/OpenCL/LICENSE $CWD/$PKG/LICENSE.OpenCL || exit 1
@@ -481,12 +254,12 @@ fi
 #cp $PREFIX/share/doc/fontconfig/COPYING $CWD/$PKG/LICENSE.fontconfig || exit 1
 #cp $PREFIX/share/doc/freetype/README $CWD/$PKG/LICENSE.freetype || exit 1
 
-if [ "$OS" == "Msys" ]; then
-  cp $PREFIX/share/doc/zlib/README $CWD/$PKG/LICENSE.zlib || exit 1
-  cp $PREFIX/share/doc/expat/COPYING $CWD/$PKG/LICENSE.expat || exit 1
-  cp $PREFIX/share/doc/fontconfig/COPYING $CWD/$PKG/LICENSE.fontconfig || exit 1
-  cp $PREFIX/share/doc/freetype/README $CWD/$PKG/LICENSE.freetype || exit 1
-fi
+#if [ "$OS" == "Msys" ]; then
+#  cp $PREFIX/share/doc/zlib/README $CWD/$PKG/LICENSE.zlib || exit 1
+#  cp $PREFIX/share/doc/expat/COPYING $CWD/$PKG/LICENSE.expat || exit 1
+#  cp $PREFIX/share/doc/fontconfig/COPYING $CWD/$PKG/LICENSE.fontconfig || exit 1
+#  cp $PREFIX/share/doc/freetype/README $CWD/$PKG/LICENSE.freetype || exit 1
+#fi
 
 # Strip and copy
 if [ "$PKGNAME" != "Arena" ]; then
